implemented the open/close and interfaces
I finished implementing a method that returns a string list of the
interfaces.  This is a combination of the cards and ports attached. They
now show up in the drop down box. I also implemented the open/close
methods and a basic write method to prove it is working. It was tested
with max and is working.